pyproject.toml
[project]
name = "ctf"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.135.1",
    "flask>=3.1.3",
    "flask-cors>=6.0.2",
    "pyyaml>=6.0.3",
    "sqlmodel>=0.0.37",
]
